requirements.txt
aws-cdk-lib
constructs
aws_cdk.aws_iot_alpha
aws-cdk.aws_iot_actions_alpha
aws_cdk.aws_dynamodb
